# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/announcements": {
            "get": {
                "tags": [
                    "announcements"
                ],
                "description": "Returns all announcements",
                "operationId": "findAnnouncements",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "priority",
                                "-priority",
                                "-title",
                                "title"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filters base on the parameter passed e.g filter[label]=event",
                        "required": false,
                        "schema": {
                            "properties": {
                                "user_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "priority": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "label": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "event"
                                    }
                                },
                                "title": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "mss walimah"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success respones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Announcement"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requested filter(s) `description, are not allowed. Allowed filter(s) are `created_at` "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/announcements/{id}": {
            "get": {
                "tags": [
                    "announcements"
                ],
                "description": "Returns an announcement based on a single ID",
                "operationId": "findAnnouncementById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of annnouncement to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "single announcement response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Announcement"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Announcement": {
                "required": [
                    "title",
                    "announcement",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "announcement": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "image": {
                        "type": "boolean"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Announcement"
                }
            },
...
        },
...
    },
    "security": [
        []
    ]
}